Temporarily reverted to double to check if grad check is more consistent. It is!

+++ Retia.Genetic/IEvolvable.cs
@@ -3,7 +3,7 @@ namespace Retia.Genetic
 	public interface IEvolvable
 	{
 		double Fitness { get; set; }
-        float[] Chromosome { get; }
+        double[] Chromosome { get; }
 
 		IEvolvable Clone();
 	}
